+++ packages/native-core-macro/Cargo.toml
@@ -12,7 +12,7 @@ keywords = ["dom", "ui", "gui", "react"]
 proc-macro = true
 
 [dependencies]
-syn = { version = "1.0.11", features = ["extra-traits", "full"] }
+syn = { version = "2.0", features = ["extra-traits", "full"] }
 quote = "1.0"
 
 [dev-dependencies]
