+++ packages/native-core-macro/src/lib.rs
@@ -14,7 +14,7 @@ pub fn partial_derive_state(_: TokenStream, input: TokenStream) -> TokenStream {
     let has_create_fn = impl_block
         .items
         .iter()
-        .any(|item| matches!(item, syn::ImplItem::Method(method) if method.sig.ident == "create"));
+        .any(|item| matches!(item, syn::ImplItem::Fn(method) if method.sig.ident == "create"));
 
     let parent_dependencies = impl_block
         .items
