+++ packages/router-macro/Cargo.toml
@@ -15,7 +15,7 @@ keywords = ["dom", "ui", "gui", "react", "router"]
 proc-macro = true
 
 [dependencies]
-syn = { version = "1.0.11", features = ["extra-traits", "full"] }
+syn = { version = "2.0", features = ["extra-traits", "full"] }
 quote = "1.0"
 proc-macro2 = "1.0.56"
 slab = "0.4"
