+++ packages/rsx-rosetta/Cargo.toml
@@ -18,7 +18,7 @@ dioxus-rsx = { workspace = true }
 html_parser.workspace = true
 proc-macro2 = "1.0.49"
 quote = "1.0.23"
-syn = { version = "1.0.107", features = ["full"] }
+syn = { version = "2.0", features = ["full"] }
 convert_case = "0.5.0"
 
 # [features]
