+++ packages/rsx/Cargo.toml
@@ -14,8 +14,12 @@ keywords = ["dom", "ui", "gui", "react"]
 [dependencies]
 proc-macro2 = { version = "1.0", features = ["span-locations"] }
 dioxus-core = { workspace = true}
-syn = { version = "1.0", features = ["full", "extra-traits"] }
+syn = { version = "2.0", features = ["full", "extra-traits"] }
 quote = { version = "1.0" }
-serde = { version = "1.0", features = ["derive"] }
-internment = "0.7.0"
-krates = "0.12.6"
+serde = { version = "1.0", features = ["derive"], optional = true }
+internment = { version = "0.7.0", optional = true }
+krates = { version = "0.12.6", optional = true }
+
+[features]
+hot_reload = ["krates", "internment"]
+serde = ["dep:serde"]
